move dataset constructor serialization tests to kernel tests
