+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -68,24 +68,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "dataset_constructor_serialization_test",
-    size = "medium",
-    srcs = ["dataset_constructor_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python:sparse_tensor",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "//third_party/py/numpy",
-    ],
-)
-
-tf_py_test(
     name = "parallel_map_dataset_serialization_test",
     size = "medium",
     srcs = ["parallel_map_dataset_serialization_test.py"],
