+++ tensorflow/python/data/kernel_tests/from_tensor_slices_test.py
@@ -22,6 +22,7 @@ import collections
 from absl.testing import parameterized
 import numpy as np
 
+from tensorflow.python.data.kernel_tests import checkpoint_test_base
 from tensorflow.python.data.kernel_tests import test_base
 from tensorflow.python.data.ops import dataset_ops
 from tensorflow.python.framework import combinations
@@ -292,22 +293,25 @@ class FromTensorSlicesTest(test_base.DatasetTestBase, parameterized.TestCase):
                      dataset_ops.get_legacy_output_types(dataset))
     self.assertDatasetProduces(dataset, expected_output)
 
-  @combinations.generate(
-      combinations.times(test_base.default_test_combinations(),
-                         combinations.combine(depth=[1, 2, 3])))
-  def testDatasetInputSerialization(self, depth):
-    dataset = dataset_ops.Dataset.range(100)
-    for _ in range(depth):
-      dataset = [dataset, dataset]
-    dataset = dataset_ops.Dataset.from_tensor_slices(dataset)
-    for _ in range(depth - 1):
-      dataset = dataset.unbatch()
-    dataset = dataset.flat_map(lambda x: x)
-    dataset = self.graphRoundTrip(dataset)
-    expected = list(range(100)) + list(range(100))
-    for _ in range(depth - 1):
-      expected = expected + expected
-    self.assertDatasetProduces(dataset, expected)
+class FromTensorSlicesCheckpointTest(checkpoint_test_base.CheckpointTestBase,
+                                     parameterized.TestCase):
+
+  def _build_tensor_slices_dataset(self, components):
+    return dataset_ops.Dataset.from_tensor_slices(components)
+
+  @combinations.generate(test_base.default_test_combinations())
+  def testFromTensorSlicesCore(self):
+    # Equal length components
+    components = (np.tile(np.array([[1], [2], [3], [4]]), 20),
+                  np.tile(np.array([[12], [13], [14], [15]]), 22),
+                  np.array([37.0, 38.0, 39.0, 40.0]))
+
+    dict_components = {"foo": [1, 2, 3], "bar": [[4.0], [5.0], [6.0]]}
+
+    self.run_core_tests(lambda: self._build_tensor_slices_dataset(components),
+                        4)
+    self.run_core_tests(
+        lambda: self._build_tensor_slices_dataset(dict_components), 3)
 
 
 if __name__ == "__main__":
